# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "RentNow API",
    "description": "API REST para la gestión de arrendamientos, propiedades, contratos y pagos.",
    "version": "1.0.0",
    "contact": {
      "name": "RentNow Support",
      "email": "support@rentnow.app",
      "url": "https://rentnow.app"
    }
  },
  "servers": [
    { "url": "https://rentnow.app", "description": "Producción" },
    { "url": "http://localhost:3000", "description": "Desarrollo" }
  ],
  "paths": {
...
    "/api/admin/organizations": {
      "get": {
        "summary": "Listar organizaciones",
        "tags": ["Admin"],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Lista de organizaciones" } }
      }
    },
    "/api/dashboard/summary": {
      "get": {
        "summary": "Resumen del dashboard (leads, riesgos, KPIs)",
        "tags": ["Dashboard"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "tenantId", "in": "query", "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": { "200": { "description": "Resumen completo del dashboard" } }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT de Supabase Auth"
      }
    }
  }
}